requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
streamlit
requests
pandas
facebook_business
openai
stripe
